backend/session-service/requirements.txt
flask
flask_cors
kafka-python
python-dotenv